requirements.txt
streamlit
chromadb
huggingface-hub
pypdf
python-dotenv